Move merchant most revenue controller
